pyproject.toml
[project]
name = "asteroids"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pygame-ce>=2.5.6",
]

[tool.ruff]
select = ["E", "F", "I"]  
unfixable = ["F401", "F841"]
fix = true